feat(service): implement full service layer + fix type-conversion compile errors
- Add service/exception: VisitNotFoundException, InvalidVisitStateException,
  AIProcessingException, PatientNotFoundException, DuplicatePatientException,
  PortalTokenExpiredException, DocumentNotFoundException
- Add service/dto: VisitCreateRequest, VisitUpdateRequest, VisitSummaryData,
  PatientPortalDTO, DashboardStatsDTO, PatientCreateRequest, PatientSearchResult,
  ChatRequest
- Add service/visit: VisitService, VisitQueryService, VisitStatusMachine
- Add service/ai: AIService, AIPromptBuilder, AISummaryParser,
  EscalationDetector, PatientChatService
- Add service/language: LanguageService (12 Indian languages, RTL Urdu)
- Add service/patient: PatientService
- Add service/doctor: DashboardStatsService (Redis-cached dashboard metrics)
- Add service/notification: NotificationService, MedicineReminderFactory (IST-aware)
- Add service/event: VisitProcessedEvent, VisitSentEvent, VisitEventListener
- Add service/document: DocumentService, OCRService
- Add domain/visit: VisitStatus enum
- Add security/config: AsyncConfig (aiProcessingExecutor, auditExecutor,
  notificationExecutor thread pools)

Bug fixes:
- NotificationRepository: doctorId param Long -> UUID (DailyDigestScheduler)
- VisitService: PatientNotFoundException pass UUID directly, not .toString()
- DashboardStatsService: explicit (int) cast on countFailedNotificationsSince
- FollowUpScheduler: LocalDate.atStartOfDay() for ISTTimeUtil.format()
- SchedulerHealthMonitor: remove unused StringRedisTemplate import

+++ backend/src/main/java/com/shifa/service/doctor/DashboardStatsService.java
@@ -28,17 +28,20 @@ public class DashboardStatsService {
         LocalDate monthStart = today.withDayOfMonth(1);
 
         return DashboardStatsDTO.builder()
-                .todayVisits((int)visitRepository.countByDoctorUserIdAndVisitDate(doctorId, today))
-                .todayFollowUps((int)visitRepository.countFollowUpsByDoctorAndDate(doctorId, today))
-                .monthVisits((int)visitRepository.countByDoctorUserIdAndVisitDateAfter(doctorId, monthStart))
-                .monthNewPatients((int)visitRepository.countNewPatientsByDoctorSince(doctorId, monthStart))
-                .pendingAiProcessing((int)visitRepository.countByDoctorUserIdAndStatus(doctorId, VisitStatus.NOTES_TAKEN))
-                .pendingSendToPatient((int)visitRepository.countByDoctorUserIdAndStatus(doctorId, VisitStatus.REVIEWED))
-                .failedNotifications(notificationRepository.countFailedNotificationsSince(weekAgo))
-                .totalPatients((int)visitRepository.countDistinctPatientsByDoctorUserId(doctorId))
-                .totalVisits((int)visitRepository.countByDoctorUserIdAndDeletedFalse(doctorId))
+                .todayVisits((int) visitRepository.countByDoctorUserIdAndVisitDate(doctorId, today))
+                .todayFollowUps((int) visitRepository.countFollowUpsByDoctorAndDate(doctorId, today))
+                .monthVisits((int) visitRepository.countByDoctorUserIdAndVisitDateAfter(doctorId, monthStart))
+                .monthNewPatients((int) visitRepository.countNewPatientsByDoctorSince(doctorId, monthStart))
+                .pendingAiProcessing(
+                        (int) visitRepository.countByDoctorUserIdAndStatus(doctorId, VisitStatus.NOTES_TAKEN))
+                .pendingSendToPatient(
+                        (int) visitRepository.countByDoctorUserIdAndStatus(doctorId, VisitStatus.REVIEWED))
+                .failedNotifications((int) notificationRepository.countFailedNotificationsSince(weekAgo))
+                .totalPatients((int) visitRepository.countDistinctPatientsByDoctorUserId(doctorId))
+                .totalVisits((int) visitRepository.countByDoctorUserIdAndDeletedFalse(doctorId))
                 .whatsappReadRate(calculateReadRate(doctorId))
-                .topLanguage(visitRepository.findTopPatientLanguageByDoctorUserId(doctorId).stream().findFirst().map(arr -> (String)arr[0]).orElse("hi"))
+                .topLanguage(visitRepository.findTopPatientLanguageByDoctorUserId(doctorId).stream().findFirst()
+                        .map(arr -> (String) arr[0]).orElse("hi"))
                 .build();
     }
